// src/app/shared/models/tab-tag/tab-side-menu.model.ts
export class TabSideMenuModel {
    tagId: string;
    tagNo: string;
    description: string;
    noOfItrCompleted: number;
    noOfItr: number;
    noOfPunchTypeACompleted: number;
    noOfPunchTypeA: number;
    noOfPunchTypeBCompleted: number;
    noOfPunchTypeB: number;
    noOfPunchTypeCCompleted: number;
    noOfPunchTypeC: number;
    noOfChangeActive: number;
    noOfChangeCompleted: number;
    noOfPreservationComplete: number;
    noOfPreservationDue: number;
    noOfPreservationOverdue: number;
}